eval/similarity/clip/requirements.txt
fastapi
uvicorn[standard]
python-multipart
torch>=2.7.1
transformers
Pillow
sentencepiece